Turn _007::Runtime::Builtins into a function
"Sometimes, the elegant implementation is just a function. Not a method.
Not a class. Not a framework. Just a function."
    -- John Carmack,
       https://twitter.com/ID_AA_Carmack/status/53512300451201024

This turned out nice. As a side effect, I noticed that we don't
need to pass the whole runtime into the builtins, just input and
output (for `say` and `prompt`).

We also inject the opscope, so that the runtime can be in control of it.
This felt preferable to exposing a second function and returning
the opscope from it.