Make infix:<~~> a lot less tight
Judging by the commit history, it seems it was just added at the
end of the list of infixes, which made it the tightest infix
operator.

This was probably not intended. It means, for example, that the
quite reasonable

    say("a" ~ "b" ~~ Str)

would not give back `1` (truthy), but result in a type error,
because it'd parse as `"a" ~ ("b" ~~ Str)`, and you can't concat
a string and a boolean.

After some thinking, I decided to put it among the comparison
operators. It is a comparison, kind of.

Added a test to assert that the precedence of infix:<~~> is
indeed this.