Introduce Val::Bool.
This was a very straightforward change. A testament to something
being very right in the architecture of 007.

This commit consists of

* adding grammar and action method support for "True" and "False"
* adding a Q::Literal::Bool type
* adding a Val::Bool type (and making sure it stringifies)
* making sure `equal-value` DTRT with Val::Bool
* removing some manual Bool->Int conversions in builtins

The rest was a matter of convincing the test suite to have the
appropriate expectations when a bunch of operators and methods
no longer give back 0 and 1, but False and True instead.

Some tests had to go. For example, this is no longer allowed,
since it relies on the intermediate value being an Int:

    say(1 < 2 <= 1);
    say(0 <= 0 < 1);

    say(1 <= 2 >= 2);
    say(2 >= 2 <= 1);

    say(1 >= 2 > 2);
    say(2 > 2 >= 1);

Likewise, there's no way to rescue this test that I can think
of:

    say(!0 * 7);

(Its purpose was to assert that `prefix:<!>` binds tighter than
`infix:<*>`. But now it fails since you can't multiply a Bool
by an Int.)

Closes #157.