[bin/007] Introduce a --backend flag
In anticipation of wanting to compile to JavaScript. See #166.

It may seem odd that we're passing in `$runtime` to the backend
closures, since the goal is to *not* use the runtime when we
compile to JS. The reason we need to do this is that the runtime
is used in two places for the default parser/runtime combo:
first in the parser, and then for running the code. Since we want
the same runtime instead of two unrelated ones, we pass it in.