Add code style test to prevent `.made` method
The `.made` method reached popularity and semi-ousted its
synonym `.ast` in the Perl 6 world. I think "made" is a
particularly bad choice of name.

It's a past participle, like "sung", "raised" or "wept".
Either of these I wouldn't be surprised to see in an `if`
statement:

    if $soufflé.made {
        ...
    }

The other possible reading I can think of is that the
`Match` object itself made something (which should then
be supplied as a parameter to the method). But that's
not what happened: *we* made something (by calling `make`)
and gave it to the `Match` object.

There's no *noun* "made". You don't ask the `Match` object
for its "made". I realize you're meant to read it as
"the thing that `make` put in there before", or
"the thing that was made". But to me it doesn't naturally
read like that.