Make anon subs stringify to <sub ()>
This looked nicer when I saw it, rather than the previous
`<sub (anon)()>`, where the parentheses clash a little bit.

Maybe there are some other situations where "(anon)" would be
nicer -- for example in a stack trace? -- but here I prefer just
the empty string. In some sense, that's a much more telling
name.

There was no test breakage, so I also added a test for the
new behavior.