Make sure X::TypeCheck :expects is a type
If it's a literal string (or any Str value, really), the expected
type will be shown as "Str", when the intent was to show that
string *describing* the type. This is because the method in
X::TypeCheck that decides what to show in the "expected" slot
looks like this:

    method expectedn() {
        (try $!got.^name eq $!expected.^name
          ?? $!expected.perl
          !! $!expected.^name
        ) // "?"
    }

In our case, we'll always hit the `!!` branch, and thus always
show the type of the string ("Str"), not the string itself.

In two of the three cases found, it wasn't really possible to
pick one type, because the description said "str or int". But
we have to pick one type, because that's how the exception works.

Also added a test to make sure this doesn't recur.