Refactor rename s/Val::None/Val::NoneType/g
Also realized that in a few (two) places, we do exact testing
against NONE, rather than type matching against NoneType.
Changed those, because even though it *should* be airtight
that we use NONE all over the place, it's not worth it to
take any chances.

So the rule from now on is: use NONE wherever, but typecheck
with ~~ Val::NoneType.

I may or may not add a code-style test to enforce that we
don't ===-check against NONE.