Simplify implementation of builtin say()
I'm pretty sure that extra line of convoluted code is some kind
of artifact of our getting more and more elaborate stringification.
There was a point when arrays were a little bit special in that
regard. The fact that the test suite doesn't complain when I
remove it is good enough for me.