Exit str(), enter prefix:<~>
It was really tempting to also loosen the requirement that only
strings can be concatenated, not (say) numbers. That would've
helped against some cases of this nature:

    say("Moving " ~ diskname ~ " from pile " ~ ~(from + 1) ~ " to pile " ~ ~(to + 1) ~ "...");

And we could've written that simply as

    say("Moving " ~ diskname ~ " from pile " ~ (from + 1) ~ " to pile " ~ (to + 1) ~ "...");

Which is slightly less, you know... tilde-y.

But I decided against it in the end. Firstly because it's out of
scope of what I'm fixing right now. Secondly because Python is
actually that strict, requiring you to make it a str type before
you concatenate. Thirdly because it would raise the question of
whether infix:<+> ought to similarly numify its operands, making
the change even bigger.

We might still decide to auto-coerce in the end. But not in this
commit.

Closes #187.