Do precedence lookup on the full name, not just op
Name lookup used to be on things like `==`, rather than `infix:==`.
Now it's the latter.

This is actually alright for infixes, since they sit in a precedence
space all by themselves. It is *not* OK for prefixes and postfixes,
and problems manifest as soon as you have both (say) `prefix:<&>`
and `postfix:<&>`; the precedence lookup would invariably find the
looser operator, even when looking for the tighter one. (That's
why the method is called `.first`, you see.)

Likely this logical error occurred because I implemented infix
precedence first, where the name-only shortcut is OK, and then
didn't re-evaluate that assumption later when implementing pre/
postfix precedence.

This is one of those rare bugfixes where the code actually becomes
a little bit simpler.

Closes #190.