Correct a regex that caused -> to be parsed wrong
This is a case where a regex silently *succeeding* (but in the
wrong way, with a partial match) causes trouble down the line.

Also the regex was wrong because it was built on old assumptions.
Commit e32e2d5c86c8eb6a6ae96c4c8efb2c72ca7916b9 made this change:

         return
             unless $identname ~~ / (< prefix infix postfix >)
    -                               ':<' (<-[>]>+) '>' /;
    +                               ':' (<-[>]>+) /;

Which removes the angle brackets around, but fails to compensate
for the fact that angle brackets inside are now allowed.

As I investigated this, I saw that there are more cases of `<-[>]>`
that are no longer based on correct assumptions. Will fix these
in later commits.

Closes #175.