Shouldn't stringify Q::Identifier anymore
Should always ask for their .name -- after we started mangling
`infix:<+>` to `infix:+` these are not the same any more and
.name has the correct version.

I could've come up with a test for this, but it'd be silly and
unrealistic.

Might be worth to get a code style test to taboo stringifying
identifiers, though.