Change find() into something safer
The `gather` block turned up weird errors in one of the test files:

    Cannot assign to a readonly variable or a value
      in code  at t/code-style/four-space-indent.t line 8
      in block <unit> at t/code-style/four-space-indent.t line 16

Tried tracking it down, but it felt like a lot of factors needed to
be present together to cause the error, so it was hard to minimize.

Changed it in all version of find(). Yes, we have a lot of versions
of find(). No, I'm not convinced they necessarily need to be
consolidated into some kind of common dependency.