Don't store the Val::Block in the frame
Good news! It turns out we didn't need to store the whole block in
the frame. Just the `outer-frame` property it carries.

It's amazing what things show up once you start to see something is
unnecessary in your model. This can be seen as a first step in
getting rid of `Val::Block` as per #205.