stop talking about Val::Block
Since nothing in 007 is creating `Val::Block` instances any more.

In all cases but one, this meant phrasing things in terms of
`Val::Sub` instead. In the remaining case, it meant simply
removing a `when` in a `given`.

`Val::Sub` still inherits from `Val::Block`, though. That's up next.

The fourth step in getting rid of `Val::Block` as per #205.