Work in progress -- messing with scopes
These two changes together (happily) make the added test pass,
fixing issue two hundred twelve. (Don't know if writing it out
would close the issue -- better not take any risks.)

Unfortunately, they also break the following test.

    {
        my $program = q:to/./;
            macro foo() {
                my x = 7;
                return quasi {
                    say(x);
                }
            }

            foo();
            .

        outputs $program, "7\n", "a variable is looked up in the quasi's environment";
    }

That test is broken by this change:

-            return $thing.new(:name($thing.name), :frame($runtime.current-frame))
+            return $thing.new(:name($thing.name), :frame($thing.frame))
                 if $thing ~~ Q::Identifier;

...which is the same line that *fixes* the new test.

I just tried putting a ternary operator
`$thing.frame === NONE ?? $runtime.current-frame !! $thing.frame` in there
instead. That unbreaks the above test, but breaks the new test. (In other
words, it seems to have zero effect relative to the baseline.)

Still, I think there has to be *some* compromise in there which will make
both tests pass. Just need to find it.

There's a total of five test failures, in these files:

* t/examples/format.t
* t/examples/name.t
* t/features/builtins/methods.t
* t/features/quasi.t (the test shown above)
* t/features/unhygienic-declarations.t

The first four are quite clearly of the same type: a variable is not found,
or found in the wrong way, because the injectile does not get the quasi's
OUTER.

The fifth has that, plus the fact that variables are not properly declared in
injectiles (so an assignment in it causes unhygienic side effects).