Last statement of quasi becomes value of quasi
That is, after the code has been injected, it has left behind a
Q::Expr::StatementListAdapter, whose sole purpose is to send on
the value of the last statement, if that happened to be an
expression. See https://github.com/masak/007/issues/7 for the
details of why we went down this path.

Closes #206.
Closes #7.