add scope-exploration
Incomplete. Had it in a gist, but it properly belongs here in the
007 repository.

My aim is to get to grips with how all the following things
interact:

- lexical scopes
- variable lookup
- stack frames
- quasiquotes
- macro arguments
- macro splicing

In a way, this list is ordered in such a way that each new entry
makes me slightly more alarmed/worried when I think about how it
interacts with all the previous ones.