Compare with None directly
Instead of doing two type checks every time.

There used to be a time when I didn't want to risk doing this in the
code base, for fear that somehow a second None instance had been
created somewhere, or would be in the future. But we're heading into
a state where the everything is set up much more deliberately, so it
feels like the risk of that is very small.