Remove test for unhygienic declarations
The original idea with this use case was to have a simple way to
signal "this `my x` should happen in the mainline scope, not in
the injectile scope". The way we signaled that was to create the
declaration as a synthetic AST.

I still think it's an important use case, and one that we should
have a simple story for, but this particular way seems less
well-advised now that I've understood that injectiles are *all*
synthetic (by virtue of being created by cloning rather than by
parsing).

So let's remove the failing test for now. We can come up with
another way to do this, I'm sure.

Closes #88.