Make Q::Block have a $.frame
In order for an injectile to have its quasi's outer, the Q::Block
generated from the quasi needs to store an outer frame to pass on
to the Q::Expr::BlockAdapter.

This is a "smallest possible change". There's also another mechanism,
magically displacing variables, which complements this one. Maybe
we only need one of these mechanisms, but

(a) the practice of magically displacing variables has been
    gradually running out of steam, and

(b) when I tried to remove it to quickly in this patch, I ended up
    killing the patient rather than healing it.

It is possible we'll end up with only one mechanism in the long term.
Or maybe they both have a role to play? We'll see.