Allow '{{{' to happen first in a statement
Closes #287.

As described in that issue, a statement starting with an unquote '{{{'
wouldn't parse properly. This was a bug, ultimately caused by the
fact that a statement starting with a '{' was deemed to *not* be an
expression statement.

The reason for this latter rule is that '{' starting a statement
introduces a *statement block*, like so:

    {
        my thing = "statement block";
        say("Hello, I'm a " ~ thing);
    }

The same disambiguation technique exists in Perl 5, Perl 6, and
JavaScript: '{' starts an object literal, except at the start of a
statement where it starts a statement block.

(Confounsing factor: I used to be able to demonstrate this as
parsing errors in Chrome Devtools, Firefox, and Node -- but all these
have lately improved their CLI parsing to introduce exceptions so
that it magically understands your intent and prefers object
literals. This also breaks some of the examples from the WAT talk.)

Fortunately, LTM offers us an out here: a '{{{' "token" is not a
type of '{' token, and so we should only disallow the former, not
the latter. It's quite unambiguously the case that '{' can not
occur at the start of an expression statement, but '{{{' *can*.