Change "sub" to "func" everywhere
I wish I knew more about the etymology, but

* subprogram
* subroutine
* routine
* procedure
* function

are all kind of the same thing. (Yes, I know they're sometimes
used differently.)

Of these, "function" is the clear winner. It's the most beginner-
friendly, and easy to understand. JavaScript has made it very
popular.

It's just it's a bit long. "function". And "fun" is too short,
and a bit frivolous.

So "func" it is.

Closes #296.