Make the builtins not take input/output
At the cost of intercepting their calls in the runtime, this change
makes the builtins() sub independent of $input and $output
parameters, and therefore independent of the input/output objects
injected into the runtime altogether.

This is a first step in making the builtins more static/constant,
in line with the vision outlined in #185.