Create builtins outside of builtins()
This moves the initialization of the builtins outside of the
builtins() function. This mostly means that we can call this
function multiple times and it will reliably return exactly
the same builtins, instead of re-creating them over and over.

Also as a necessary part of this, we expose an `opscope()`
function that hands back the separately-built opscope for the
builtins. This rids us of the (frankly a bit silly) previous
mechanism where it needed to be passed in to `builtins()`
and initialized as a side effect.