Remove or rewrite all AST-based tests
Some tests were not possible to rewrite because they were
intimately tied to the AST representation of programs. Some were
removed because they were duplicates of tests already done with
parsing the program.

Most tests could be rewritten into a smaller/simpler 007 program
instead of the old AST format.

Unfortunately when we are testing for errors, we are not testing
for error messages. I'd like to reinstate the error messages in
a later change, because it's good practice to test for them.