Remove Q::Statement::My in favor of Q::Term::My
See #279 for reasons.

In the end this was a smaller change than I feared it would be. Having made it,
I'm also more convinced (despite initial misgivings) that it's a simpler model.

(In particular, it used to be that assignment semantics resided "inside" the
`my` statement semantics. Now it's all on the outside. The `my` term, at
runtime, is completely transparent and behaves exactly like the identifier it
declares.)

It also doesn't hurt that this simpler model is also somewhat richer than the
old one. :)

Closes #279.