Change '::' to '.' as a package separator
This change charges some extra onto the already debt-ridden
'property' method in Runtime.pm6. It's overdue for some
cleanup; maybe even before one that affects the entire object
system.

In general, this change introduces technical debt here and
there, which can either be cleaned up gradually, or as part
of a bigger object system rework. Until then, we already
have the dot separator, and we've cleaned up `::` for other
things (such as a custom operator).

There are a couple of places in the language where previously
we expected "one identifier" denoting a type, but now it's a
chain of dot-separated identifiers instead; what #250 calls a
"name":

- After `new`
- After `quasi @`
- Before the `@` in unquotes

The "neat" shortcut in self-host/runtime.007 where we used
the Q hierarchy type "identifiers" as object/dictionary keys
had to go. But that was a price we had to pay at some point
anyway.

Closes #250.
Closes #298.