Introduce Val::Object, make it the top type
This re-establishes Val::Object, but this time as the kind of
empty type that goes at the top of a type hierarchy.

As a consequence, "everything is an Object", which right now is
hard-coded into the ~~ and !~~ operators.

In the fullness of time, we'll instead code it into the type
hierarchy itself, so that a special case is not needed.

Closes #144.