Limit Dicts to o["key"] syntax, not o.key
Also limit a number of methods to Dict only.

This is a natural, unavoidable separation of Dict and Object.
I think the biggest downstream client to suffer from it
syntactically is self-host/runtime.007, because now we have to
write

    Runtime.new()

as

    Runtime["new"]()

which isn't nearly as pretty.

But on the other hand, the long-term fix to that is to introduce
real classes.