Check function terms in injectiles
We didn't have function terms until
58653425c14c30736f2b8ff3439436a613712f46, and the check function
apparently never dove into them. Now that it does, it allows the
initial reported wrong line of #212 to run just fine:

    $ bin/007 -e='macro moo(x) { return quasi { (func() { my y = {{{x}}} })() }; }; say(moo(42))'
    42

I think #212 can be closed now, but going to have a look around
and try to run everything in it, and also check out #207 which I
know was affected by this, and also think through what check might
be missing.