Introduce 'export' on 'func'/'macro'/'my'
This keyword doesn't have any semantics at all yet, since we haven't
started implementing modules, not even behind a feature flag. The
idea to have this keyword now is to be able to write 007 code that's
closer to what modules will eventually look like when we do have
them. Less to change to make them work.

To be extra safe, and avoid having to break backwards compatibility
later, we make sure that if an expression is exported, then it has
a leftmost `my` to export.

Closes #404.