Add Bresenham algorithm
In its present form, it's terribly repetitive, and not a good
example of DRY code.

The goal of this branch is to explore ways, possibly macroish
ways, to condense the main loop into _one_ copy, while also not
making the code too contorted and hard to read. That is,
abstraction with zero cost in speed or in complexity.