Installing a new op removes the old of that name
A new invariant on the OpScope object is that an operator (say
`infix:+`) is only allowed to occur *once* in the lookup tables.
When a new operator is added, any operator with the same name is
removed.

An odd consequence of this is that we can end up with precedence
levels containing no operator. Although that's an odd thing, I
don't see any negative effects. In particular, since precedence
levels are not first-class and only accessible via the ops they
contain, an empty precedence level will simply be inaccessible.

To be honest, this fix should have been how it was done from the
start. We clone the OpScope object anyway, so it makes sense to
destroy old ops from the copy.

Closes #372.