Make built-ins unassignable
Before this change, we already got an error when trying to assign
to a built-in function or type. But the error was "Variable not
declared", which is confusing. The name *is* declared, just not
inside of the program text itself. But the check that led to the
error didn't consider that.

What we do now with an assignment is, if we find a variable that
hasn't been declared within the program text, we tentatively make
a (static) lookup on that variable.

- If we find something, then it's a built-in, and we throw
  X::Assignment::ReadOnly. [This is the new bit.]

- If we don't find anything, then (like before), we throw
  X::Undefined.

Closes #411.