Quarantine some of the tests
These, incidentally, are the ones that require hygiene. Previously
we provided that using the `identifier.frame` mechanism, but I'm
about to remove that on this branch and put in another one.

Before merge, these tests need to work again using the new mechanism.
But for now, switching them off like this so that the rest of the
test suite stays green.