Separate Q::Identifier and Q::Term::Identifier
The former is the general form, when we have a name in code that
shouldn't ever do any kind of value lookup, lexical or otherwise.

Common examples include dictionary keys, property names, and all
kind of definitions (`my`, `class`, `func`, etc.).

The latter is a term in an expression. Basically anything that
ought to look up a value at runtime should use this form.

This change was harder than it looked. A few corners have been
cut, and can hopefully be done better in the long term.