Exclude quasi variables from being made direct
This is the wrong fix; it concludes that a variable was declared in
the quasi iff lookup for it frame fails.

The problem is that lookup might succeed (and find a variable
outside of the quasi) but the variable might still have been
declared inside the quasi.

There's a better fix; it's probably a bit more complicated, and
must involve actually declaring those variables inside the quasi
block, including in nested blocks. I think the `interpolate`
sub will end up looking a lot more like the `check` sub.

Because this is the wrong fix, some things are still left in
quarantine. However, quite a few things moved back, too.