Restore .detach to working order
I have a feeling this method will need its own mechanism once we
get to actually remembering which block variables outside of a quasi
come from. For now, this works.

This empties out the quarantined tests file, and so we pass all the
tests we used to, and then some.