Refactor charclass transitions to be multi-range
Previously we had one transition per range. Now we can stuff all
of the ranges in the same transition.

For simplicity, all the elements of a `Transition::Ranges` are
ranges, even when the element matches only one character. In that
case, we do our best to lie creatively and pretend it's just one
character.

This refactor is in preparation for negative character classes,
(for example \W) which need to consider all their ranges at the
same time in order to match them negatively together.

The "real" solution, able to support <+[abc]-[def]> and similar
like Perl 6 does, is probably some appropriate data structure
like a skiplist. But I doubt we'll need that generality for this
spike at least.