Add a "mention" example
This file is a contrived example, meant to highlight the fact that, even though eight lines
of code are printed, only four macro calls are made.

I have no idea whether this will work or not in the current implementation. It should, and
I know of no reason why it would not, but I haven't tried it. A test needs to be added
before this can be merged.